refactor(obsidian): extract wikilinks module from obsidian converter
Moved slug_from_url, should_convert_wikilink, convert_wiki_links,
transform_and_serialize, push_event_text, extract_text_from_events
to new wikilinks.rs module.

obsidian.rs: 678 → 190 lines (-72%)
wikilinks.rs: 428 lines (new)
Shared push_event_text kept pub(crate) for obsidian reuse.
All 21 tests passing (16 wikilinks + 5 obsidian).
