+++ src/infrastructure/converter/mod.rs
@@ -4,3 +4,4 @@ pub mod html_cleaner;
 pub mod html_to_markdown;
 pub mod obsidian;
 pub mod syntax_highlight;
+pub mod wikilinks;
